requirements-base.txt
numpy>=1.26.0,<2.1.0
pandas>=2.1.0,<2.2.0
scikit-learn>=1.3.0,<1.4.0
scipy>=1.11.0,<1.12.0
xgboost>=2.0.0,<2.1.0

mlflow>=2.10.0

PyYAML>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0